@@ -4,24 +4,22 @@
   "private": true,
   "devDependencies": {
     "axios-mock-adapter": "1.16.0",
+    "@babel/core": "^7.12.9",
+    "@babel/runtime": "^7.12.5",
+    "@react-native-community/eslint-config": "^2.0.0",
+    "babel-jest": "^26.6.3",
+    "eslint": "7.14.0",
+    "jest": "^26.6.3",
+    "metro-react-native-babel-preset": "^0.64.0",
+    "react-test-renderer": "17.0.1",
     "chai": "4.2.0",
-    "eslint": "5.16.0",
-    "eslint-config-standard": "12.0.0",
-    "eslint-plugin-import": "2.17.3",
     "eslint-plugin-node": "9.1.0",
     "eslint-plugin-promise": "4.1.1",
     "eslint-plugin-standard": "4.0.0",
-    "jest": "24.8.0",
     "jest-fetch-mock": "2.1.2",
     "jest-react-native": "18.0.0",
     "jest-serializer-enzyme": "1.0.0",
     "jsdom": "15.1.1",
-    "react-addons-test-utils": "15.6.2",
-    "react-dom": "16.8.6",
-    "react-native-mock": "0.3.1",
-    "react-native-mock-render": "0.1.4",
-    "react-test-renderer": "16.8.6",
-    "reactotron-react-native": "3.6.4",
     "sinon": "7.3.2"
   },
   "scripts": {
@@ -60,25 +58,23 @@
     "@fortawesome/pro-light-svg-icons": "5.9.0",
     "@fortawesome/pro-regular-svg-icons": "5.9.0",
     "@fortawesome/pro-solid-svg-icons": "5.9.0",
-    "@react-native-community/async-storage": "1.5.0",
-    "@react-native-community/cli": "1.9.12",
-    "@react-native-community/netinfo": "^4.1.3",
-    "@react-navigation/core": "3.4.2",
-    "@react-navigation/native": "3.5.0",
-    "axios": "0.19.0",
+    "@react-native-async-storage/async-storage": "^1.15.5",
+    "@react-native-clipboard/clipboard": "^1.8.2",
+    "@react-native-community/netinfo": "^6.0.0",
+    "axios": "^0.21.1",
     "babel-preset-react-native": "4.0.1",
     "base-64": "0.1.0",
     "crypto-js": "3.1.9-1",
     "events": "3.0.0",
     "jetifier": "^1.6.3",
-    "lodash": "4.17.11",
+    "lodash": "^4.17.21",
     "metro-react-native-babel-preset": "0.54.1",
     "moment": "2.24.0",
-    "nachos-ui": "0.2.0-beta.1",
+    "nachos-ui": "^0.2.0-beta.1",
     "ndaujs": "github:ndau/ndaujs",
     "prop-types": "15.7.2",
-    "react": "16.8.6",
-    "react-native": "0.60.4",
+    "react": "17.0.1",
+    "react-native": "0.64.2",
     "react-native-autocomplete-input": "3.6.0",
     "react-native-background-fetch": "2.6.0",
     "react-native-barcode-mask": "1.0.5",
@@ -88,28 +84,28 @@
     "react-native-device-info": "2.1.2",
     "react-native-flash-message": "0.1.13",
     "react-native-fontawesome-pro": "2.1.0",
-    "react-native-fs": "2.13.3",
-    "react-native-gesture-handler": "1.3.0",
+    "react-native-fs": "^2.18.0",
+    "react-native-gesture-handler": "^1.10.3",
     "react-native-linear-gradient": "2.5.4",
     "react-native-loading-spinner-overlay": "1.0.1",
-    "react-native-mail": "3.0.7",
+    "react-native-mail": "^3.0.7",
     "react-native-modal": "^11.1.0",
     "react-native-modal-dropdown": "0.6.2",
     "react-native-permissions": "1.1.1",
     "react-native-push-notification": "github:ndau/react-native-push-notification#kp-specificgoogleversion",
-    "react-native-qrcode-svg": "5.1.2",
+    "react-native-qrcode-svg": "^6.1.1",
+    "react-native-reanimated": "^1.13.1",
     "react-native-responsive-screen": "1.2.2",
-    "react-native-securerandom": "0.3.1",
+    "react-native-safe-area-context": "^3.2.0",
+    "react-native-screens": "^3.4.0",
+    "react-native-securerandom": "^1.0.0",
     "react-native-share": "1.2.1",
     "react-native-svg": "9.5.1",
     "react-native-version-number": "0.3.6",
-    "react-navigation": "3.11.0"
-  },
-  "rnpm": {
-    "assets": [
-      "assets/fonts",
-      "./assets/fonts"
-    ]
+    "react-navigation": "^4.4.4",
+    "react-navigation-drawer": "^2.7.1",
+    "react-navigation-stack": "^2.10.4",
+    "reactotron-react-native": "^5.0.0"
   },
   "react-native": {
     "zlib": "browserify-zlib",
