@@ -37,11 +37,20 @@ class AccountLockConfirmation extends Component {
   constructor (props) {
     super(props)
     this.state = {
-      account: {},
-      wallet: {},
-      lockInformation: {},
-      accountAddressForEAI: null,
-      accountNicknameForEAI: null,
+      account: AccountStore.getAccount(),
+      wallet: WalletStore.getWallet(),
+      lockInformation: props.navigation.getParam(
+        'lockInformation',
+        null
+      ),
+      accountAddressForEAI: props.navigation.getParam(
+        'accountAddressForEAI',
+        null
+      ),
+      accountNicknameForEAI: props.navigation.getParam(
+        'accountNicknameForEAI',
+        null
+      ),
       confirmed: false,
       word: null,
       spinner: false,
@@ -50,20 +59,71 @@ class AccountLockConfirmation extends Component {
     props.navigation.addListener('didBlur', FlashNotification.hideMessage)
   }
 
-  componentWillMount = () => {
-    const account = AccountStore.getAccount()
-    const wallet = WalletStore.getWallet()
-    const lockInformation = this.props.navigation.getParam(
-      'lockInformation',
-      null
+//   componentDidMount = () => {
+//     const account = AccountStore.getAccount()
+//     const wallet = WalletStore.getWallet()
+//     const lockInformation = this.props.navigation.getParam(
+//       'lockInformation',
+//       null
+//     )
+//     const accountAddressForEAI = this.props.navigation.getParam(
+//       'accountAddressForEAI',
+//       null
+//     )
+//     const accountNicknameForEAI = this.props.navigation.getParam(
+//       'accountNicknameForEAI',
+//       null
+//     )
+
+//     this.setState({ spinner: true }, async () => {
+//       let transactionFee = ''
+//       try {
+//         Object.assign(LockTransaction.prototype, Transaction)
+//         this.lockTransaction = new LockTransaction(
+//           this.state.wallet,
+//           account,
+//           `${this.state.lockInformation.lockISO}`
+//         )
+//         await this.lockTransaction.create()
+//         await this.lockTransaction.sign()
+//         const data = await this.lockTransaction.prevalidate()
+//         transactionFee = new NdauNumber(data.fee_napu).toDetail()
+//         this.setState({
+//           spinner: false,
+//           transactionFee
+//         })
+//       } catch (error) {
+//         const err = new BlockchainAPIError(error)
+//         FlashNotification.showError(err.message, false)
+//         this.setState({
+//           spinner: false,
+//           transactionFee: '[error]'
+//         })
+//       }
+//     })
+
+//     this.setState({
+//       account,
+//       wallet,
+//       lockInformation,
+//       accountAddressForEAI,
+//       accountNicknameForEAI
+//     })
+//   }
+
+  componentDidMount () {
+    Object.assign(NotifyTransaction.prototype, Transaction)
+    this.notifyTransaction = new NotifyTransaction(
+      this.state.wallet,
+      this.state.account
     )
-    const accountAddressForEAI = this.props.navigation.getParam(
-      'accountAddressForEAI',
-      null
-    )
-    const accountNicknameForEAI = this.props.navigation.getParam(
-      'accountNicknameForEAI',
-      null
+
+    // Now make sure we send the EAI where it belongs
+    Object.assign(SetRewardsDestinationTransaction.prototype, Transaction)
+    this.setRewardsDestinationTransaction = new SetRewardsDestinationTransaction(
+      this.state.wallet,
+      this.state.account,
+      this.state.accountAddressForEAI
     )
 
     this.setState({ spinner: true }, async () => {
@@ -72,7 +132,7 @@ class AccountLockConfirmation extends Component {
         Object.assign(LockTransaction.prototype, Transaction)
         this.lockTransaction = new LockTransaction(
           this.state.wallet,
-          account,
+          this.state.account,
           `${this.state.lockInformation.lockISO}`
         )
         await this.lockTransaction.create()
@@ -92,30 +152,6 @@ class AccountLockConfirmation extends Component {
         })
       }
     })
-
-    this.setState({
-      account,
-      wallet,
-      lockInformation,
-      accountAddressForEAI,
-      accountNicknameForEAI
-    })
-  }
-
-  componentDidMount () {
-    Object.assign(NotifyTransaction.prototype, Transaction)
-    this.notifyTransaction = new NotifyTransaction(
-      this.state.wallet,
-      this.state.account
-    )
-
-    // Now make sure we send the EAI where it belongs
-    Object.assign(SetRewardsDestinationTransaction.prototype, Transaction)
-    this.setRewardsDestinationTransaction = new SetRewardsDestinationTransaction(
-      this.state.wallet,
-      this.state.account,
-      this.state.accountAddressForEAI
-    )
   }
 
   _lock = async () => {
