@@ -9,7 +9,8 @@
  */
 
 import React from 'react'
-import { createStackNavigator, createDrawerNavigator } from 'react-navigation'
+import { createDrawerNavigator } from 'react-navigation-drawer'
+import { createStackNavigator } from 'react-navigation-stack'
 import Dashboard from '../screens/Dashboard'
 import AppDrawer from './AppDrawer'
 
@@ -17,7 +18,7 @@ const DashboardScreen = ({ navigation }) => (
   <Dashboard navigation={navigation} />
 )
 DashboardScreen.navigationOptions = ({ navigation }) => ({
-  header: null
+    headerShown: false
 })
 
 const DashboardStack = createStackNavigator({
