@@ -9,13 +9,14 @@
  */
 
 import React from 'react'
-import { createStackNavigator, createDrawerNavigator } from 'react-navigation'
+import { createDrawerNavigator } from 'react-navigation-drawer'
+import { createStackNavigator } from 'react-navigation-stack'
 import Settings from '../screens/Settings'
 import AppDrawer from './AppDrawer'
 
 const SettingsScreen = ({ navigation }) => <Settings navigation={navigation} />
 SettingsScreen.navigationOptions = ({ navigation }) => ({
-  header: null
+    headerShown: false
 })
 
 const SettingsStack = createStackNavigator({
